backend/pyproject.toml
[project]
name = "zerofootprint-backend"
version = "0.1.0"
description = "ZeroFootprint backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "sqlmodel>=0.0.16",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.1",
  "python-dotenv>=1.0",
  "cryptography>=42.0",
  "pyjwt>=2.8",
  "httpx>=0.27",
  "redis>=5.0",
  "rq>=1.16",
  "google-auth>=2.29",
  "google-auth-oauthlib>=1.2",
  "google-api-python-client>=2.125",
]

[project.optional-dependencies]
test = [
  "pytest>=8.0",
  "pytest-mock>=3.12",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
